% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orb.R
\name{BufferPool}
\alias{BufferPool}
\title{Buffer Pool for ORB}
\description{
Manages a pool of reusable CDR buffers
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BufferPool-new}{\code{BufferPool$new()}}
\item \href{#method-BufferPool-acquire_buf}{\code{BufferPool$acquire_buf()}}
\item \href{#method-BufferPool-release_buf}{\code{BufferPool$release_buf()}}
\item \href{#method-BufferPool-outstanding_buffer_count}{\code{BufferPool$outstanding_buffer_count()}}
\item \href{#method-BufferPool-clone}{\code{BufferPool$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BufferPool-new"></a>}}
\if{latex}{\out{\hypertarget{method-BufferPool-new}{}}}
\subsection{Method \code{new()}}{
Create buffer pool
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BufferPool$new(orb, buffer_size = 65536L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{orb}}{ORB instance}

\item{\code{buffer_size}}{Size of each buffer}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BufferPool-acquire_buf"></a>}}
\if{latex}{\out{\hypertarget{method-BufferPool-acquire_buf}{}}}
\subsection{Method \code{acquire_buf()}}{
Acquire a buffer from the pool
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BufferPool$acquire_buf()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
CDR buffer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BufferPool-release_buf"></a>}}
\if{latex}{\out{\hypertarget{method-BufferPool-release_buf}{}}}
\subsection{Method \code{release_buf()}}{
Release a buffer back to the pool
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BufferPool$release_buf(buf)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{buf}}{Buffer to release}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BufferPool-outstanding_buffer_count"></a>}}
\if{latex}{\out{\hypertarget{method-BufferPool-outstanding_buffer_count}{}}}
\subsection{Method \code{outstanding_buffer_count()}}{
Get count of outstanding buffers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BufferPool$outstanding_buffer_count()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Integer count
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BufferPool-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BufferPool-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BufferPool$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
