% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_buffer.R
\name{CdrBuffer}
\alias{CdrBuffer}
\title{CDR Buffer for Binary Wire Format}
\description{
Buffer class for reading and writing the binary wire format used by the ORB.
This implements the Common Data Representation (CDR) protocol for marshalling
data between R and the SLC process.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CdrBuffer-new}{\code{CdrBuffer$new()}}
\item \href{#method-CdrBuffer-clear}{\code{CdrBuffer$clear()}}
\item \href{#method-CdrBuffer-flip}{\code{CdrBuffer$flip()}}
\item \href{#method-CdrBuffer-position}{\code{CdrBuffer$position()}}
\item \href{#method-CdrBuffer-set_position}{\code{CdrBuffer$set_position()}}
\item \href{#method-CdrBuffer-limit}{\code{CdrBuffer$limit()}}
\item \href{#method-CdrBuffer-set_limit}{\code{CdrBuffer$set_limit()}}
\item \href{#method-CdrBuffer-capacity}{\code{CdrBuffer$capacity()}}
\item \href{#method-CdrBuffer-reserve}{\code{CdrBuffer$reserve()}}
\item \href{#method-CdrBuffer-remaining}{\code{CdrBuffer$remaining()}}
\item \href{#method-CdrBuffer-buffer}{\code{CdrBuffer$buffer()}}
\item \href{#method-CdrBuffer-write_boolean}{\code{CdrBuffer$write_boolean()}}
\item \href{#method-CdrBuffer-write_byte}{\code{CdrBuffer$write_byte()}}
\item \href{#method-CdrBuffer-write_short}{\code{CdrBuffer$write_short()}}
\item \href{#method-CdrBuffer-write_int}{\code{CdrBuffer$write_int()}}
\item \href{#method-CdrBuffer-write_long}{\code{CdrBuffer$write_long()}}
\item \href{#method-CdrBuffer-write_float}{\code{CdrBuffer$write_float()}}
\item \href{#method-CdrBuffer-write_double}{\code{CdrBuffer$write_double()}}
\item \href{#method-CdrBuffer-write_string}{\code{CdrBuffer$write_string()}}
\item \href{#method-CdrBuffer-write_object}{\code{CdrBuffer$write_object()}}
\item \href{#method-CdrBuffer-read_boolean}{\code{CdrBuffer$read_boolean()}}
\item \href{#method-CdrBuffer-read_byte}{\code{CdrBuffer$read_byte()}}
\item \href{#method-CdrBuffer-read_short}{\code{CdrBuffer$read_short()}}
\item \href{#method-CdrBuffer-read_int}{\code{CdrBuffer$read_int()}}
\item \href{#method-CdrBuffer-read_long}{\code{CdrBuffer$read_long()}}
\item \href{#method-CdrBuffer-read_float}{\code{CdrBuffer$read_float()}}
\item \href{#method-CdrBuffer-read_double}{\code{CdrBuffer$read_double()}}
\item \href{#method-CdrBuffer-read_string}{\code{CdrBuffer$read_string()}}
\item \href{#method-CdrBuffer-read_object}{\code{CdrBuffer$read_object()}}
\item \href{#method-CdrBuffer-clone}{\code{CdrBuffer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CdrBuffer-new"></a>}}
\if{latex}{\out{\hypertarget{method-CdrBuffer-new}{}}}
\subsection{Method \code{new()}}{
Create a new CDR buffer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CdrBuffer$new(orb, initial_capacity = 4096L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{orb}}{The ORB instance}

\item{\code{initial_capacity}}{Initial buffer capacity in bytes}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CdrBuffer-clear"></a>}}
\if{latex}{\out{\hypertarget{method-CdrBuffer-clear}{}}}
\subsection{Method \code{clear()}}{
Clear the buffer ready for read
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CdrBuffer$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CdrBuffer-flip"></a>}}
\if{latex}{\out{\hypertarget{method-CdrBuffer-flip}{}}}
\subsection{Method \code{flip()}}{
Flip buffer from write to read mode
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CdrBuffer$flip()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CdrBuffer-position"></a>}}
\if{latex}{\out{\hypertarget{method-CdrBuffer-position}{}}}
\subsection{Method \code{position()}}{
Get current position
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CdrBuffer$position()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CdrBuffer-set_position"></a>}}
\if{latex}{\out{\hypertarget{method-CdrBuffer-set_position}{}}}
\subsection{Method \code{set_position()}}{
Set position
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CdrBuffer$set_position(pos)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pos}}{New position}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CdrBuffer-limit"></a>}}
\if{latex}{\out{\hypertarget{method-CdrBuffer-limit}{}}}
\subsection{Method \code{limit()}}{
Get current limit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CdrBuffer$limit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CdrBuffer-set_limit"></a>}}
\if{latex}{\out{\hypertarget{method-CdrBuffer-set_limit}{}}}
\subsection{Method \code{set_limit()}}{
Set limit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CdrBuffer$set_limit(limit)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{limit}}{New limit}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CdrBuffer-capacity"></a>}}
\if{latex}{\out{\hypertarget{method-CdrBuffer-capacity}{}}}
\subsection{Method \code{capacity()}}{
Get buffer capacity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CdrBuffer$capacity()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CdrBuffer-reserve"></a>}}
\if{latex}{\out{\hypertarget{method-CdrBuffer-reserve}{}}}
\subsection{Method \code{reserve()}}{
Reserve capacity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CdrBuffer$reserve(new_cap)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_cap}}{New capacity}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CdrBuffer-remaining"></a>}}
\if{latex}{\out{\hypertarget{method-CdrBuffer-remaining}{}}}
\subsection{Method \code{remaining()}}{
Get remaining bytes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CdrBuffer$remaining()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CdrBuffer-buffer"></a>}}
\if{latex}{\out{\hypertarget{method-CdrBuffer-buffer}{}}}
\subsection{Method \code{buffer()}}{
Get buffer as raw vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CdrBuffer$buffer()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CdrBuffer-write_boolean"></a>}}
\if{latex}{\out{\hypertarget{method-CdrBuffer-write_boolean}{}}}
\subsection{Method \code{write_boolean()}}{
Write a boolean
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CdrBuffer$write_boolean(b)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{b}}{Boolean value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CdrBuffer-write_byte"></a>}}
\if{latex}{\out{\hypertarget{method-CdrBuffer-write_byte}{}}}
\subsection{Method \code{write_byte()}}{
Write a byte
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CdrBuffer$write_byte(b)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{b}}{Byte value (0-255)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CdrBuffer-write_short"></a>}}
\if{latex}{\out{\hypertarget{method-CdrBuffer-write_short}{}}}
\subsection{Method \code{write_short()}}{
Write a short (2 bytes)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CdrBuffer$write_short(s)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{s}}{Short integer value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CdrBuffer-write_int"></a>}}
\if{latex}{\out{\hypertarget{method-CdrBuffer-write_int}{}}}
\subsection{Method \code{write_int()}}{
Write an integer (4 bytes)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CdrBuffer$write_int(i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{Integer value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CdrBuffer-write_long"></a>}}
\if{latex}{\out{\hypertarget{method-CdrBuffer-write_long}{}}}
\subsection{Method \code{write_long()}}{
Write a long (8 bytes)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CdrBuffer$write_long(l)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{l}}{Long integer value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CdrBuffer-write_float"></a>}}
\if{latex}{\out{\hypertarget{method-CdrBuffer-write_float}{}}}
\subsection{Method \code{write_float()}}{
Write a float (4 bytes)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CdrBuffer$write_float(f)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f}}{Float value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CdrBuffer-write_double"></a>}}
\if{latex}{\out{\hypertarget{method-CdrBuffer-write_double}{}}}
\subsection{Method \code{write_double()}}{
Write a double (8 bytes)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CdrBuffer$write_double(d)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d}}{Double value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CdrBuffer-write_string"></a>}}
\if{latex}{\out{\hypertarget{method-CdrBuffer-write_string}{}}}
\subsection{Method \code{write_string()}}{
Write a string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CdrBuffer$write_string(s)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{s}}{String value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CdrBuffer-write_object"></a>}}
\if{latex}{\out{\hypertarget{method-CdrBuffer-write_object}{}}}
\subsection{Method \code{write_object()}}{
Write an object reference
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CdrBuffer$write_object(obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{Object instance}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CdrBuffer-read_boolean"></a>}}
\if{latex}{\out{\hypertarget{method-CdrBuffer-read_boolean}{}}}
\subsection{Method \code{read_boolean()}}{
Read a boolean
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CdrBuffer$read_boolean()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Boolean value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CdrBuffer-read_byte"></a>}}
\if{latex}{\out{\hypertarget{method-CdrBuffer-read_byte}{}}}
\subsection{Method \code{read_byte()}}{
Read a byte
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CdrBuffer$read_byte()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Byte value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CdrBuffer-read_short"></a>}}
\if{latex}{\out{\hypertarget{method-CdrBuffer-read_short}{}}}
\subsection{Method \code{read_short()}}{
Read a short
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CdrBuffer$read_short()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Short integer value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CdrBuffer-read_int"></a>}}
\if{latex}{\out{\hypertarget{method-CdrBuffer-read_int}{}}}
\subsection{Method \code{read_int()}}{
Read an integer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CdrBuffer$read_int()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Integer value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CdrBuffer-read_long"></a>}}
\if{latex}{\out{\hypertarget{method-CdrBuffer-read_long}{}}}
\subsection{Method \code{read_long()}}{
Read a long
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CdrBuffer$read_long()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Long integer value (as numeric)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CdrBuffer-read_float"></a>}}
\if{latex}{\out{\hypertarget{method-CdrBuffer-read_float}{}}}
\subsection{Method \code{read_float()}}{
Read a float
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CdrBuffer$read_float()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Float value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CdrBuffer-read_double"></a>}}
\if{latex}{\out{\hypertarget{method-CdrBuffer-read_double}{}}}
\subsection{Method \code{read_double()}}{
Read a double
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CdrBuffer$read_double()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Double value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CdrBuffer-read_string"></a>}}
\if{latex}{\out{\hypertarget{method-CdrBuffer-read_string}{}}}
\subsection{Method \code{read_string()}}{
Read a string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CdrBuffer$read_string()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CdrBuffer-read_object"></a>}}
\if{latex}{\out{\hypertarget{method-CdrBuffer-read_object}{}}}
\subsection{Method \code{read_object()}}{
Read an object reference
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CdrBuffer$read_object(type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Object class}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Object instance or NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CdrBuffer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CdrBuffer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CdrBuffer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
