% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto-integration.R
\name{is_quarto_context}
\alias{is_quarto_context}
\title{Check if we're in a Quarto rendering context}
\usage{
is_quarto_context()
}
\value{
Logical indicating if in Quarto context
}
\description{
Check if we're in a Quarto rendering context
}
