% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{ObjectAdapter}
\alias{ObjectAdapter}
\title{Object Adapter}
\description{
Manages servant objects in the ORB
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ObjectAdapter-new}{\code{ObjectAdapter$new()}}
\item \href{#method-ObjectAdapter-add}{\code{ObjectAdapter$add()}}
\item \href{#method-ObjectAdapter-add_with_uuid}{\code{ObjectAdapter$add_with_uuid()}}
\item \href{#method-ObjectAdapter-remove}{\code{ObjectAdapter$remove()}}
\item \href{#method-ObjectAdapter-id_to_servant}{\code{ObjectAdapter$id_to_servant()}}
\item \href{#method-ObjectAdapter-dispatch}{\code{ObjectAdapter$dispatch()}}
\item \href{#method-ObjectAdapter-clone}{\code{ObjectAdapter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObjectAdapter-new"></a>}}
\if{latex}{\out{\hypertarget{method-ObjectAdapter-new}{}}}
\subsection{Method \code{new()}}{
Create object adapter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObjectAdapter$new(orb)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{orb}}{The ORB instance}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObjectAdapter-add"></a>}}
\if{latex}{\out{\hypertarget{method-ObjectAdapter-add}{}}}
\subsection{Method \code{add()}}{
Add a servant with specific ID
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObjectAdapter$add(servant, object_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{servant}}{Servant object}

\item{\code{object_id}}{Object ID}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Object reference
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObjectAdapter-add_with_uuid"></a>}}
\if{latex}{\out{\hypertarget{method-ObjectAdapter-add_with_uuid}{}}}
\subsection{Method \code{add_with_uuid()}}{
Add a servant with UUID
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObjectAdapter$add_with_uuid(servant)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{servant}}{Servant object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Object reference
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObjectAdapter-remove"></a>}}
\if{latex}{\out{\hypertarget{method-ObjectAdapter-remove}{}}}
\subsection{Method \code{remove()}}{
Remove a servant
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObjectAdapter$remove(object_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object_id}}{Object ID or servant}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObjectAdapter-id_to_servant"></a>}}
\if{latex}{\out{\hypertarget{method-ObjectAdapter-id_to_servant}{}}}
\subsection{Method \code{id_to_servant()}}{
Get servant by ID
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObjectAdapter$id_to_servant(object_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object_id}}{Object ID}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Servant object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObjectAdapter-dispatch"></a>}}
\if{latex}{\out{\hypertarget{method-ObjectAdapter-dispatch}{}}}
\subsection{Method \code{dispatch()}}{
Dispatch a method call
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObjectAdapter$dispatch(object_id, operation, msgbuf, replybuf)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object_id}}{Target object ID}

\item{\code{operation}}{Operation name}

\item{\code{msgbuf}}{Message buffer}

\item{\code{replybuf}}{Reply buffer}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Reply status
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObjectAdapter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ObjectAdapter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObjectAdapter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
