% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orb.R
\name{Orb}
\alias{Orb}
\title{Object Request Broker (ORB)}
\description{
Main ORB class for managing communication with SLC
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Orb-new}{\code{Orb$new()}}
\item \href{#method-Orb-object_adapter}{\code{Orb$object_adapter()}}
\item \href{#method-Orb-string_to_object}{\code{Orb$string_to_object()}}
\item \href{#method-Orb-shutdown}{\code{Orb$shutdown()}}
\item \href{#method-Orb-wait_for_reply}{\code{Orb$wait_for_reply()}}
\item \href{#method-Orb-acquire_buf}{\code{Orb$acquire_buf()}}
\item \href{#method-Orb-release_buf}{\code{Orb$release_buf()}}
\item \href{#method-Orb-send}{\code{Orb$send()}}
\item \href{#method-Orb-recv}{\code{Orb$recv()}}
\item \href{#method-Orb-generate_request_id}{\code{Orb$generate_request_id()}}
\item \href{#method-Orb-clone}{\code{Orb$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Orb-new"></a>}}
\if{latex}{\out{\hypertarget{method-Orb-new}{}}}
\subsection{Method \code{new()}}{
Create ORB instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Orb$new(conn, process_handle = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{conn}}{Connection object}

\item{\code{process_handle}}{Optional process handle for liveness checking}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Orb-object_adapter"></a>}}
\if{latex}{\out{\hypertarget{method-Orb-object_adapter}{}}}
\subsection{Method \code{object_adapter()}}{
Get object adapter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Orb$object_adapter()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ObjectAdapter instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Orb-string_to_object"></a>}}
\if{latex}{\out{\hypertarget{method-Orb-string_to_object}{}}}
\subsection{Method \code{string_to_object()}}{
Convert string to object reference
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Orb$string_to_object(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Object ID string}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Object reference
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Orb-shutdown"></a>}}
\if{latex}{\out{\hypertarget{method-Orb-shutdown}{}}}
\subsection{Method \code{shutdown()}}{
Shutdown the ORB
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Orb$shutdown()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Orb-wait_for_reply"></a>}}
\if{latex}{\out{\hypertarget{method-Orb-wait_for_reply}{}}}
\subsection{Method \code{wait_for_reply()}}{
Wait for reply to a request
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Orb$wait_for_reply(waiter, request_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{waiter}}{Waiter object}

\item{\code{request_id}}{Request ID}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List with buffer and header
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Orb-acquire_buf"></a>}}
\if{latex}{\out{\hypertarget{method-Orb-acquire_buf}{}}}
\subsection{Method \code{acquire_buf()}}{
Acquire a buffer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Orb$acquire_buf()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
CDR buffer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Orb-release_buf"></a>}}
\if{latex}{\out{\hypertarget{method-Orb-release_buf}{}}}
\subsection{Method \code{release_buf()}}{
Release a buffer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Orb$release_buf(buf)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{buf}}{Buffer to release}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Orb-send"></a>}}
\if{latex}{\out{\hypertarget{method-Orb-send}{}}}
\subsection{Method \code{send()}}{
Send data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Orb$send(buf)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{buf}}{Buffer to send}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Orb-recv"></a>}}
\if{latex}{\out{\hypertarget{method-Orb-recv}{}}}
\subsection{Method \code{recv()}}{
Receive data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Orb$recv(buf)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{buf}}{Buffer to receive into}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Orb-generate_request_id"></a>}}
\if{latex}{\out{\hypertarget{method-Orb-generate_request_id}{}}}
\subsection{Method \code{generate_request_id()}}{
Generate request ID
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Orb$generate_request_id()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Integer request ID
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Orb-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Orb-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Orb$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
