% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{OrbObject}
\alias{OrbObject}
\title{ORB Object Reference}
\description{
Represents a reference to a remote object in the ORB
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrbObject-new}{\code{OrbObject$new()}}
\item \href{#method-OrbObject-orb}{\code{OrbObject$orb()}}
\item \href{#method-OrbObject-identity}{\code{OrbObject$identity()}}
\item \href{#method-OrbObject-is_a}{\code{OrbObject$is_a()}}
\item \href{#method-OrbObject-request}{\code{OrbObject$request()}}
\item \href{#method-OrbObject-invoke_oneway}{\code{OrbObject$invoke_oneway()}}
\item \href{#method-OrbObject-invoke}{\code{OrbObject$invoke()}}
\item \href{#method-OrbObject-release_buf}{\code{OrbObject$release_buf()}}
\item \href{#method-OrbObject-clone}{\code{OrbObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrbObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrbObject-new}{}}}
\subsection{Method \code{new()}}{
Create object reference
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrbObject$new(orb, id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{orb}}{The ORB instance}

\item{\code{id}}{Object identity string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrbObject-orb"></a>}}
\if{latex}{\out{\hypertarget{method-OrbObject-orb}{}}}
\subsection{Method \code{orb()}}{
Get the ORB
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrbObject$orb()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ORB instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrbObject-identity"></a>}}
\if{latex}{\out{\hypertarget{method-OrbObject-identity}{}}}
\subsection{Method \code{identity()}}{
Get object identity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrbObject$identity()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Identity string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrbObject-is_a"></a>}}
\if{latex}{\out{\hypertarget{method-OrbObject-is_a}{}}}
\subsection{Method \code{is_a()}}{
Check if object is of given type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrbObject$is_a(type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Type string}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Boolean
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrbObject-request"></a>}}
\if{latex}{\out{\hypertarget{method-OrbObject-request}{}}}
\subsection{Method \code{request()}}{
Create a request buffer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrbObject$request(operation)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{operation}}{Operation name}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
CDR buffer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrbObject-invoke_oneway"></a>}}
\if{latex}{\out{\hypertarget{method-OrbObject-invoke_oneway}{}}}
\subsection{Method \code{invoke_oneway()}}{
Invoke a oneway operation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrbObject$invoke_oneway(msgbuf)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msgbuf}}{Message buffer}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrbObject-invoke"></a>}}
\if{latex}{\out{\hypertarget{method-OrbObject-invoke}{}}}
\subsection{Method \code{invoke()}}{
Invoke an operation and wait for reply
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrbObject$invoke(msgbuf)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msgbuf}}{Message buffer}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Reply buffer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrbObject-release_buf"></a>}}
\if{latex}{\out{\hypertarget{method-OrbObject-release_buf}{}}}
\subsection{Method \code{release_buf()}}{
Release a buffer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrbObject$release_buf(buff)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{buff}}{Buffer to release}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrbObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrbObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrbObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
