% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read_slc_data}
\alias{read_slc_data}
\title{Read SLC Dataset into R}
\usage{
read_slc_data(dataset_name, connection = NULL)
}
\arguments{
\item{dataset_name}{Name of the SLC dataset}

\item{connection}{SLC connection object}
}
\value{
R data.frame
}
\description{
Read SLC Dataset into R
}
\examples{
\dontrun{
# Initialize connection
conn <- slc_init()

# First write some data
write_slc_data(mtcars, "mtcars_dataset", conn)

# Then read it back
mtcars_copy <- read_slc_data("mtcars_dataset", conn)

# Compare dimensions
dim(mtcars)
dim(mtcars_copy)

# Compare elememt by element
mtcars_copy == mtcars
}
}
