% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto-integration.R
\name{setup_slc_quarto}
\alias{setup_slc_quarto}
\title{Setup SLC for Quarto (called automatically when library is loaded)}
\usage{
setup_slc_quarto()
}
\description{
Setup SLC for Quarto (called automatically when library is loaded)
}
