% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto-integration.R
\name{slc_resource_path}
\alias{slc_resource_path}
\title{Get path to SLC Quarto resources}
\usage{
slc_resource_path(file)
}
\arguments{
\item{file}{The resource file name}
}
\value{
Path to the resource file
}
\description{
Get path to SLC Quarto resources
}
