% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{slc_utils}
\alias{slc_utils}
\alias{saslog}
\alias{slc_r_exception}
\title{Utility functions for SLC operations}
\usage{
saslog(msg)

slc_r_exception(exc_file, error)
}
\arguments{
\item{msg}{Character string with message to log}

\item{exc_file}{Character string with exception file path}

\item{error}{Error object}
}
\description{
Utility functions for logging and error handling in SLC operations.
}
