% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slcr-package.R
\docType{package}
\name{slcR-package}
\alias{slcR}
\alias{slcR-package}
\title{slcR: R Interface to Altair SLC}
\description{
Provides an R interface to Altair SLC, enabling users to execute SLC code and transfer data between R and SLC environments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sol-eng/slcR}
  \item \url{https://sol-eng.github.io/slcR/}
  \item Report bugs at \url{https://github.com/sol-eng/slcR/issues}
}

}
\author{
\strong{Maintainer}: Michael Mayer \email{michael.mayer@posit.co}

}
\keyword{internal}
