% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpslink.R
\name{WpsDataset}
\alias{WpsDataset}
\title{WPS Dataset Object}
\description{
Stub class for WPS Dataset
}
\section{Super class}{
\code{\link[slcR:OrbObject]{slcR::OrbObject}} -> \code{WpsDataset}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WpsDataset-close}{\code{WpsDataset$close()}}
\item \href{#method-WpsDataset-get_nobs}{\code{WpsDataset$get_nobs()}}
\item \href{#method-WpsDataset-get_nvars}{\code{WpsDataset$get_nvars()}}
\item \href{#method-WpsDataset-clone}{\code{WpsDataset$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="identity"><a href='../../slcR/html/OrbObject.html#method-OrbObject-identity'><code>slcR::OrbObject$identity()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="initialize"><a href='../../slcR/html/OrbObject.html#method-OrbObject-initialize'><code>slcR::OrbObject$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="invoke"><a href='../../slcR/html/OrbObject.html#method-OrbObject-invoke'><code>slcR::OrbObject$invoke()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="invoke_oneway"><a href='../../slcR/html/OrbObject.html#method-OrbObject-invoke_oneway'><code>slcR::OrbObject$invoke_oneway()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="is_a"><a href='../../slcR/html/OrbObject.html#method-OrbObject-is_a'><code>slcR::OrbObject$is_a()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="orb"><a href='../../slcR/html/OrbObject.html#method-OrbObject-orb'><code>slcR::OrbObject$orb()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="release_buf"><a href='../../slcR/html/OrbObject.html#method-OrbObject-release_buf'><code>slcR::OrbObject$release_buf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="request"><a href='../../slcR/html/OrbObject.html#method-OrbObject-request'><code>slcR::OrbObject$request()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsDataset-close"></a>}}
\if{latex}{\out{\hypertarget{method-WpsDataset-close}{}}}
\subsection{Method \code{close()}}{
Close dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsDataset$close()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsDataset-get_nobs"></a>}}
\if{latex}{\out{\hypertarget{method-WpsDataset-get_nobs}{}}}
\subsection{Method \code{get_nobs()}}{
Get number of observations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsDataset$get_nobs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsDataset-get_nvars"></a>}}
\if{latex}{\out{\hypertarget{method-WpsDataset-get_nvars}{}}}
\subsection{Method \code{get_nvars()}}{
Get number of variables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsDataset$get_nvars()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsDataset-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WpsDataset-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsDataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
