% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpslink.R
\name{WpsLibref}
\alias{WpsLibref}
\title{WPS Libref Object}
\description{
Stub class for WPS Library reference
}
\section{Super class}{
\code{\link[slcR:OrbObject]{slcR::OrbObject}} -> \code{WpsLibref}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WpsLibref-get_name}{\code{WpsLibref$get_name()}}
\item \href{#method-WpsLibref-get_member_names}{\code{WpsLibref$get_member_names()}}
\item \href{#method-WpsLibref-open_dataset}{\code{WpsLibref$open_dataset()}}
\item \href{#method-WpsLibref-create_dataset}{\code{WpsLibref$create_dataset()}}
\item \href{#method-WpsLibref-clone}{\code{WpsLibref$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="identity"><a href='../../slcR/html/OrbObject.html#method-OrbObject-identity'><code>slcR::OrbObject$identity()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="initialize"><a href='../../slcR/html/OrbObject.html#method-OrbObject-initialize'><code>slcR::OrbObject$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="invoke"><a href='../../slcR/html/OrbObject.html#method-OrbObject-invoke'><code>slcR::OrbObject$invoke()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="invoke_oneway"><a href='../../slcR/html/OrbObject.html#method-OrbObject-invoke_oneway'><code>slcR::OrbObject$invoke_oneway()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="is_a"><a href='../../slcR/html/OrbObject.html#method-OrbObject-is_a'><code>slcR::OrbObject$is_a()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="orb"><a href='../../slcR/html/OrbObject.html#method-OrbObject-orb'><code>slcR::OrbObject$orb()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="release_buf"><a href='../../slcR/html/OrbObject.html#method-OrbObject-release_buf'><code>slcR::OrbObject$release_buf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="request"><a href='../../slcR/html/OrbObject.html#method-OrbObject-request'><code>slcR::OrbObject$request()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsLibref-get_name"></a>}}
\if{latex}{\out{\hypertarget{method-WpsLibref-get_name}{}}}
\subsection{Method \code{get_name()}}{
Get library name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsLibref$get_name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsLibref-get_member_names"></a>}}
\if{latex}{\out{\hypertarget{method-WpsLibref-get_member_names}{}}}
\subsection{Method \code{get_member_names()}}{
Get member names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsLibref$get_member_names()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character vector of member names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsLibref-open_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-WpsLibref-open_dataset}{}}}
\subsection{Method \code{open_dataset()}}{
Open dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsLibref$open_dataset(name, mode = "INPUT")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Dataset name}

\item{\code{mode}}{Open mode (default "INPUT")}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
WpsDataset object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsLibref-create_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-WpsLibref-create_dataset}{}}}
\subsection{Method \code{create_dataset()}}{
Create dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsLibref$create_dataset(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Dataset name}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
WpsDataset object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsLibref-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WpsLibref-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsLibref$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
