% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpslink.R
\name{WpsListingFile}
\alias{WpsListingFile}
\title{WPS Listing File Object}
\description{
Stub class for WPS Listing File
}
\section{Super class}{
\code{\link[slcR:OrbObject]{slcR::OrbObject}} -> \code{WpsListingFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WpsListingFile-get_page_count}{\code{WpsListingFile$get_page_count()}}
\item \href{#method-WpsListingFile-get_contents}{\code{WpsListingFile$get_contents()}}
\item \href{#method-WpsListingFile-get_page}{\code{WpsListingFile$get_page()}}
\item \href{#method-WpsListingFile-close}{\code{WpsListingFile$close()}}
\item \href{#method-WpsListingFile-clone}{\code{WpsListingFile$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="identity"><a href='../../slcR/html/OrbObject.html#method-OrbObject-identity'><code>slcR::OrbObject$identity()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="initialize"><a href='../../slcR/html/OrbObject.html#method-OrbObject-initialize'><code>slcR::OrbObject$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="invoke"><a href='../../slcR/html/OrbObject.html#method-OrbObject-invoke'><code>slcR::OrbObject$invoke()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="invoke_oneway"><a href='../../slcR/html/OrbObject.html#method-OrbObject-invoke_oneway'><code>slcR::OrbObject$invoke_oneway()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="is_a"><a href='../../slcR/html/OrbObject.html#method-OrbObject-is_a'><code>slcR::OrbObject$is_a()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="orb"><a href='../../slcR/html/OrbObject.html#method-OrbObject-orb'><code>slcR::OrbObject$orb()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="release_buf"><a href='../../slcR/html/OrbObject.html#method-OrbObject-release_buf'><code>slcR::OrbObject$release_buf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="request"><a href='../../slcR/html/OrbObject.html#method-OrbObject-request'><code>slcR::OrbObject$request()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsListingFile-get_page_count"></a>}}
\if{latex}{\out{\hypertarget{method-WpsListingFile-get_page_count}{}}}
\subsection{Method \code{get_page_count()}}{
Get page count
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsListingFile$get_page_count()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Integer number of pages
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsListingFile-get_contents"></a>}}
\if{latex}{\out{\hypertarget{method-WpsListingFile-get_contents}{}}}
\subsection{Method \code{get_contents()}}{
Get listing contents by reading all pages
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsListingFile$get_contents()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String containing all listing output
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsListingFile-get_page"></a>}}
\if{latex}{\out{\hypertarget{method-WpsListingFile-get_page}{}}}
\subsection{Method \code{get_page()}}{
Get a specific page
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsListingFile$get_page(pagenum)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pagenum}}{Page number (0-indexed)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List with exists (logical) and page (list with lines)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsListingFile-close"></a>}}
\if{latex}{\out{\hypertarget{method-WpsListingFile-close}{}}}
\subsection{Method \code{close()}}{
Close listing file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsListingFile$close()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsListingFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WpsListingFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsListingFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
