% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpslink.R
\name{WpsLogFile}
\alias{WpsLogFile}
\title{WPS Log File Object}
\description{
Stub class for WPS Log File
}
\section{Super class}{
\code{\link[slcR:OrbObject]{slcR::OrbObject}} -> \code{WpsLogFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WpsLogFile-get_line_count}{\code{WpsLogFile$get_line_count()}}
\item \href{#method-WpsLogFile-get_contents}{\code{WpsLogFile$get_contents()}}
\item \href{#method-WpsLogFile-get_lines}{\code{WpsLogFile$get_lines()}}
\item \href{#method-WpsLogFile-close}{\code{WpsLogFile$close()}}
\item \href{#method-WpsLogFile-clone}{\code{WpsLogFile$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="identity"><a href='../../slcR/html/OrbObject.html#method-OrbObject-identity'><code>slcR::OrbObject$identity()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="initialize"><a href='../../slcR/html/OrbObject.html#method-OrbObject-initialize'><code>slcR::OrbObject$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="invoke"><a href='../../slcR/html/OrbObject.html#method-OrbObject-invoke'><code>slcR::OrbObject$invoke()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="invoke_oneway"><a href='../../slcR/html/OrbObject.html#method-OrbObject-invoke_oneway'><code>slcR::OrbObject$invoke_oneway()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="is_a"><a href='../../slcR/html/OrbObject.html#method-OrbObject-is_a'><code>slcR::OrbObject$is_a()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="orb"><a href='../../slcR/html/OrbObject.html#method-OrbObject-orb'><code>slcR::OrbObject$orb()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="release_buf"><a href='../../slcR/html/OrbObject.html#method-OrbObject-release_buf'><code>slcR::OrbObject$release_buf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="request"><a href='../../slcR/html/OrbObject.html#method-OrbObject-request'><code>slcR::OrbObject$request()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsLogFile-get_line_count"></a>}}
\if{latex}{\out{\hypertarget{method-WpsLogFile-get_line_count}{}}}
\subsection{Method \code{get_line_count()}}{
Get line count
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsLogFile$get_line_count()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Integer number of lines
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsLogFile-get_contents"></a>}}
\if{latex}{\out{\hypertarget{method-WpsLogFile-get_contents}{}}}
\subsection{Method \code{get_contents()}}{
Get log contents by reading all lines
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsLogFile$get_contents()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String containing all log output
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsLogFile-get_lines"></a>}}
\if{latex}{\out{\hypertarget{method-WpsLogFile-get_lines}{}}}
\subsection{Method \code{get_lines()}}{
Get a range of lines
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsLogFile$get_lines(first, max_count)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{first}}{First line index (0-indexed)}

\item{\code{max_count}}{Maximum number of lines to retrieve}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List with result code and lines
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsLogFile-close"></a>}}
\if{latex}{\out{\hypertarget{method-WpsLogFile-close}{}}}
\subsection{Method \code{close()}}{
Close log file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsLogFile$close()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsLogFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WpsLogFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsLogFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
