% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpslink.R
\name{WpsServer}
\alias{WpsServer}
\title{WPS Server Object}
\description{
Stub class for WPS Server interface
}
\section{Super class}{
\code{\link[slcR:OrbObject]{slcR::OrbObject}} -> \code{WpsServer}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WpsServer-create_session}{\code{WpsServer$create_session()}}
\item \href{#method-WpsServer-shutdown_server}{\code{WpsServer$shutdown_server()}}
\item \href{#method-WpsServer-get_dns_name}{\code{WpsServer$get_dns_name()}}
\item \href{#method-WpsServer-get_os_name}{\code{WpsServer$get_os_name()}}
\item \href{#method-WpsServer-clone}{\code{WpsServer$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="identity"><a href='../../slcR/html/OrbObject.html#method-OrbObject-identity'><code>slcR::OrbObject$identity()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="initialize"><a href='../../slcR/html/OrbObject.html#method-OrbObject-initialize'><code>slcR::OrbObject$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="invoke"><a href='../../slcR/html/OrbObject.html#method-OrbObject-invoke'><code>slcR::OrbObject$invoke()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="invoke_oneway"><a href='../../slcR/html/OrbObject.html#method-OrbObject-invoke_oneway'><code>slcR::OrbObject$invoke_oneway()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="is_a"><a href='../../slcR/html/OrbObject.html#method-OrbObject-is_a'><code>slcR::OrbObject$is_a()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="orb"><a href='../../slcR/html/OrbObject.html#method-OrbObject-orb'><code>slcR::OrbObject$orb()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="release_buf"><a href='../../slcR/html/OrbObject.html#method-OrbObject-release_buf'><code>slcR::OrbObject$release_buf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="request"><a href='../../slcR/html/OrbObject.html#method-OrbObject-request'><code>slcR::OrbObject$request()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsServer-create_session"></a>}}
\if{latex}{\out{\hypertarget{method-WpsServer-create_session}{}}}
\subsection{Method \code{create_session()}}{
Create session
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsServer$create_session()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
WpsSession object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsServer-shutdown_server"></a>}}
\if{latex}{\out{\hypertarget{method-WpsServer-shutdown_server}{}}}
\subsection{Method \code{shutdown_server()}}{
Shutdown server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsServer$shutdown_server()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsServer-get_dns_name"></a>}}
\if{latex}{\out{\hypertarget{method-WpsServer-get_dns_name}{}}}
\subsection{Method \code{get_dns_name()}}{
Get DNS name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsServer$get_dns_name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsServer-get_os_name"></a>}}
\if{latex}{\out{\hypertarget{method-WpsServer-get_os_name}{}}}
\subsection{Method \code{get_os_name()}}{
Get OS name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsServer$get_os_name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsServer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WpsServer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsServer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
