% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpslink.R
\name{WpsSession}
\alias{WpsSession}
\title{WPS Session Object}
\description{
Stub class for WPS Session interface
}
\section{Super class}{
\code{\link[slcR:OrbObject]{slcR::OrbObject}} -> \code{WpsSession}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WpsSession-init}{\code{WpsSession$init()}}
\item \href{#method-WpsSession-init_with_options}{\code{WpsSession$init_with_options()}}
\item \href{#method-WpsSession-submit}{\code{WpsSession$submit()}}
\item \href{#method-WpsSession-get_libref}{\code{WpsSession$get_libref()}}
\item \href{#method-WpsSession-assign_libref}{\code{WpsSession$assign_libref()}}
\item \href{#method-WpsSession-get_macro_variable}{\code{WpsSession$get_macro_variable()}}
\item \href{#method-WpsSession-set_macro_variable}{\code{WpsSession$set_macro_variable()}}
\item \href{#method-WpsSession-clear_listing_file}{\code{WpsSession$clear_listing_file()}}
\item \href{#method-WpsSession-get_log_file}{\code{WpsSession$get_log_file()}}
\item \href{#method-WpsSession-get_listing_file}{\code{WpsSession$get_listing_file()}}
\item \href{#method-WpsSession-clone}{\code{WpsSession$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="identity"><a href='../../slcR/html/OrbObject.html#method-OrbObject-identity'><code>slcR::OrbObject$identity()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="initialize"><a href='../../slcR/html/OrbObject.html#method-OrbObject-initialize'><code>slcR::OrbObject$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="invoke"><a href='../../slcR/html/OrbObject.html#method-OrbObject-invoke'><code>slcR::OrbObject$invoke()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="invoke_oneway"><a href='../../slcR/html/OrbObject.html#method-OrbObject-invoke_oneway'><code>slcR::OrbObject$invoke_oneway()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="is_a"><a href='../../slcR/html/OrbObject.html#method-OrbObject-is_a'><code>slcR::OrbObject$is_a()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="orb"><a href='../../slcR/html/OrbObject.html#method-OrbObject-orb'><code>slcR::OrbObject$orb()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="release_buf"><a href='../../slcR/html/OrbObject.html#method-OrbObject-release_buf'><code>slcR::OrbObject$release_buf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="slcR" data-topic="OrbObject" data-id="request"><a href='../../slcR/html/OrbObject.html#method-OrbObject-request'><code>slcR::OrbObject$request()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsSession-init"></a>}}
\if{latex}{\out{\hypertarget{method-WpsSession-init}{}}}
\subsection{Method \code{init()}}{
Initialize session
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsSession$init()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsSession-init_with_options"></a>}}
\if{latex}{\out{\hypertarget{method-WpsSession-init_with_options}{}}}
\subsection{Method \code{init_with_options()}}{
Initialize session with options
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsSession$init_with_options(options)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{options}}{Named list of system options}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsSession-submit"></a>}}
\if{latex}{\out{\hypertarget{method-WpsSession-submit}{}}}
\subsection{Method \code{submit()}}{
Submit SAS code
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsSession$submit(code)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{code}}{SAS code string}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Integer return code
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsSession-get_libref"></a>}}
\if{latex}{\out{\hypertarget{method-WpsSession-get_libref}{}}}
\subsection{Method \code{get_libref()}}{
Get library reference
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsSession$get_libref(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Library name}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
WpsLibref object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsSession-assign_libref"></a>}}
\if{latex}{\out{\hypertarget{method-WpsSession-assign_libref}{}}}
\subsection{Method \code{assign_libref()}}{
Assign library
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsSession$assign_libref(name, path, engine = "WPD")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Library name}

\item{\code{path}}{Library path}

\item{\code{engine}}{Engine name (default "WPD")}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
WpsLibref object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsSession-get_macro_variable"></a>}}
\if{latex}{\out{\hypertarget{method-WpsSession-get_macro_variable}{}}}
\subsection{Method \code{get_macro_variable()}}{
Get macro variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsSession$get_macro_variable(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Variable name}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
String value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsSession-set_macro_variable"></a>}}
\if{latex}{\out{\hypertarget{method-WpsSession-set_macro_variable}{}}}
\subsection{Method \code{set_macro_variable()}}{
Set macro variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsSession$set_macro_variable(name, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Variable name}

\item{\code{value}}{Variable value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsSession-clear_listing_file"></a>}}
\if{latex}{\out{\hypertarget{method-WpsSession-clear_listing_file}{}}}
\subsection{Method \code{clear_listing_file()}}{
Clear listing file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsSession$clear_listing_file()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsSession-get_log_file"></a>}}
\if{latex}{\out{\hypertarget{method-WpsSession-get_log_file}{}}}
\subsection{Method \code{get_log_file()}}{
Get log file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsSession$get_log_file()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
WpsLogFile object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsSession-get_listing_file"></a>}}
\if{latex}{\out{\hypertarget{method-WpsSession-get_listing_file}{}}}
\subsection{Method \code{get_listing_file()}}{
Get listing file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsSession$get_listing_file()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
WpsListingFile object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WpsSession-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WpsSession-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WpsSession$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
