% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read_slc_data}
\alias{read_slc_data}
\title{Read SLC Dataset into R}
\usage{
read_slc_data(dataset_name, connection = NULL)
}
\arguments{
\item{dataset_name}{Name of the SLC dataset}

\item{connection}{SLC connection object}
}
\value{
R data.frame
}
\description{
Read SLC Dataset into R
}
