% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{slc_submit}
\alias{slc_submit}
\title{Submit SLC Code}
\usage{
slc_submit(code, connection = NULL)
}
\arguments{
\item{code}{Character string containing SLC code}

\item{connection}{SLC connection object}
}
\value{
Results from SLC execution
}
\description{
Submit SLC Code
}
