% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{write_slc_data}
\alias{write_slc_data}
\title{Convert R DataFrame to SLC Dataset}
\usage{
write_slc_data(data, dataset_name, connection = NULL)
}
\arguments{
\item{data}{R data.frame to convert}

\item{dataset_name}{Name for the SLC dataset}

\item{connection}{SLC connection object}
}
\value{
Invisible NULL
}
\description{
Convert R DataFrame to SLC Dataset
}
